+++ src/index.ts
@@ -2,8 +2,9 @@ import type { DiscoverConfig } from './lib/config'
 
 import { discoverConfigSchemaWithDefaults } from './lib/config'
 import { generateJsonSchema } from './lib/generate/json'
+import { generateMarkdownSchema } from './lib/generate/markdown'
 import { generateOpenApiSchema } from './lib/generate/openapi'
-import { generateTypeScriptTypes } from './lib/generate/typescript'
+import { generateTypescriptTypes } from './lib/generate/typescript'
 import { generateZodSchemas } from './lib/generate/zod'
 import { parseSchemas } from './lib/parse/schema'
 import { probeEndpoints } from './lib/probe'
@@ -57,10 +58,11 @@ export default async function discover(config: DiscoverConfig) {
   const schemaResults = await parseSchemas(probeResults, parsedConfig)
 
   const openapiResult = await generateOpenApiSchema(schemaResults, parsedConfig)
+  const typescriptResults = await generateTypescriptTypes(schemaResults, parsedConfig)
+  const zodResults = await generateZodSchemas(schemaResults, parsedConfig)
+  const jsonResults = await generateJsonSchema(schemaResults, parsedConfig)
 
-  await generateTypeScriptTypes(openapiResult, parsedConfig)
-  await generateZodSchemas(schemaResults, parsedConfig)
-  await generateJsonSchema(schemaResults, parsedConfig)
+  await generateMarkdownSchema(schemaResults, openapiResult, typescriptResults, zodResults, jsonResults, parsedConfig)
 
   const totalTime = Math.ceil(performance.now() - startTime)
   const totalProbingTime = Math.ceil(probingCompletedTime - startTime)
@@ -70,13 +72,6 @@ export default async function discover(config: DiscoverConfig) {
   parsedConfig.logger.success(`Discovery completed in ${totalTime} ms (probing took ${totalProbingTime} ms)`)
 }
 
-export type {
-  DiscoverConfig,
-  DiscoverHooks,
-  HookResult,
-  HttpHeaders,
-  HttpMethod,
-  ProbeConfig,
-  ProbeResult,
-  SchemaResult,
-} from './lib/config'
+export * from './lib/config'
+
+export type * from './types'