+++ src/lib/generate/typescript.ts
@@ -1,36 +1,51 @@
+import type { SchemaResult } from '../../types'
 import type { ParsedDiscoverConfig } from '../config'
 
 import { mkdir, writeFile } from 'node:fs/promises'
 import { joinURL } from 'ufo'
 
+import { resolveTypeName } from '../../helpers/path'
+import { buildAST } from './ast/builder'
+import { writeAST } from './ast/writer'
+
+function generateFile(name: string, types: string) {
+  return `export type ${name} = ${types};`
+}
+
 /**
- * Generate TypeScript types from OpenAPI schema result
+ * Generate TypeScript types from parsed schema results
  *
- * @param openApiResult OpenAPI schema result as JSON string
+ * @param schemaResults Array of schema results
  * @param config Parsed discovery configuration
  */
-export async function generateTypeScriptTypes(openApiResult: string | undefined, config: ParsedDiscoverConfig) {
-  if (!config.generate?.typescript || !openApiResult) {
+export async function generateTypescriptTypes(schemaResults: SchemaResult[], config: ParsedDiscoverConfig) {
+  if (!config.generate?.typescript) {
     return
   }
 
-  return import('openapi-typescript').then(async (openapiTS) => {
-    let openapiTSOptions = {}
+  const components = schemaResults.map(result => ({
+    name: resolveTypeName(joinURL(config.baseUrl ?? '', result.path)),
+    path: result.path,
+    method: result.method,
+    output: result.schema,
+  }))
 
-    if (typeof config.generate?.typescript === 'object') {
-      openapiTSOptions = config.generate.typescript
+  const schemas = await Promise.all(components.map(async ({ output, ...rest }) => {
+    await config.hooks.callHook('typescript:generate', config, rest.method, rest.name, output)
+
+    return {
+      output: writeAST(buildAST(output), { lang: 'typescript' }),
+      ...rest,
     }
+  }))
 
-    await config.hooks.callHook('typescript:generate', config, openapiTSOptions)
+  await Promise.all(schemas.map(async ({ method }) =>
+    mkdir(joinURL(config.outputDir, 'typescript', method), { recursive: true })))
 
-    const ast = await openapiTS.default(openApiResult, openapiTSOptions)
-    const result = openapiTS.astToString(ast)
+  await Promise.all(schemas.map(async ({ name, method, output }) =>
+    writeFile(joinURL(config.outputDir, 'typescript', method, `${name}.ts`), generateFile(name, output))))
 
-    await config.hooks.callHook('typescript:generated', config, result)
+  await config.hooks.callHook('typescript:generated', config, schemas)
 
-    await mkdir(joinURL(config.outputDir, 'typescript'), { recursive: true })
-      .then(() => writeFile(joinURL(config.outputDir, 'typescript', 'types.d.ts'), result))
-  }).catch((error) => {
-    throw new Error('openapi-typescript is required to generate TypeScript types.\nYou can install it with: npm install openapi-typescript', { cause: error })
-  })
+  return schemas
 }